requirement.txt
streamlit==1.42.0
pillow==11.1.0